Started updating JoinUtils;
